terminación de las rutas de BookDelete y bookUpdate

+++ api/src/routes/controllers/books/bookDelete.js
@@ -23,11 +23,13 @@ const bookDelete = async (req, res, next) => {
       }
       const book = await Book.findById(id)
 
-      await Comment.deleteMany({ _id: { $in: book.comment } })
+      if (parseInt(book.user) === parseInt(decodedToken.id)) {
+        await Comment.deleteMany({ _id: { $in: book.comment } })
 
-      const deletedBook = await Book.findByIdAndDelete(id)
+        const deletedBook = await Book.findByIdAndDelete(id)
 
-      res.status(200).json(`The workbook  ${deletedBook.title} was deleted`)
+        res.status(200).json(`The workbook  ${deletedBook.title} was deleted`)
+      }
     }
   } catch (error) {
     next(error)
